Refactor

+++ Heatmap/XMLParser.cs
@@ -1,12 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml;
-using System.IO;
-using System.Drawing;
-
 namespace Heatmap
 {
     /// <summary>
@@ -14,14 +8,22 @@ namespace Heatmap
     /// </summary>
     public class XMLParser
     {
-        List<KeyValuePair<int, int>> points;
+        List<KeyValuePair<double, double>> points;
         string filePath;
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="path">path to catalog with gpx files</param>
         public XMLParser (string path)
         {
-            points = new List<KeyValuePair<int, int>>();
+            points = new List<KeyValuePair<double, double>>();
             filePath = path;
         }
-        public List<KeyValuePair<int, int>> GetAllPoints()
+        /// <summary>
+        /// get all points from single gpx file
+        /// </summary>
+        /// <returns>list of points</returns>
+        public List<KeyValuePair<double, double>> GetAllPoints()
         {
             XmlDocument document = new XmlDocument();
             document.Load(filePath);
@@ -29,10 +31,9 @@ namespace Heatmap
             {
                 double lattitude = Convert.ToDouble(node.Attributes[0].Value.Replace(".", ","));
                 double longitude = Convert.ToDouble(node.Attributes[1].Value.Replace(".", ","));
-                points.Add(new KeyValuePair<int, int>((int)lattitude, (int)longitude));
+                points.Add(new KeyValuePair<double, double>(lattitude, longitude));
             }
             return points;
         }
-
     }
 }
